Add request handler stream examples
Compare standard lambda/graal lambda doing the same super simple hello
world task.

Need to better benchmarking, but it seems like the Graal does much
better on initial warm up, but slightly worse on recurring hot runs.

Though this is so simple and the deviations are so small it probably
isn't significant, especially for a more real world lambda with JSON
parsing and some IO / compute.